Added support for TRNG test driver with PSA config
For the option when a TRNG accerator is present in the driver
the user will have the ability to use that instead of the
CTR_DRBG and AES libraries. This set of changes puts in the basic
functionality to enable those options for the user. It still needs
to have the basic trng driver functionality filled in.

Signed-off-by: John Durkop <john.durkop@fermatsoftware.com>